[DAD-776] refactor: scrap, memo, user로 dto, repository, service 폴더 구조 변경
