+++ src/main/java/com/forever/dadamda/controller/ScrapController.java
@@ -3,16 +3,16 @@ package com.forever.dadamda.controller;
 import com.forever.dadamda.dto.ApiResponse;
 import com.forever.dadamda.dto.scrap.CreateScrapRequest;
 import com.forever.dadamda.dto.scrap.CreateScrapResponse;
-import com.forever.dadamda.dto.scrap.GetArticleCountResponse;
-import com.forever.dadamda.dto.scrap.GetArticleResponse;
-import com.forever.dadamda.dto.scrap.GetOtherCountResponse;
-import com.forever.dadamda.dto.scrap.GetOtherResponse;
-import com.forever.dadamda.dto.scrap.GetProductCountResponse;
-import com.forever.dadamda.dto.scrap.GetProductResponse;
+import com.forever.dadamda.dto.scrap.article.GetArticleCountResponse;
+import com.forever.dadamda.dto.scrap.article.GetArticleResponse;
+import com.forever.dadamda.dto.scrap.other.GetOtherCountResponse;
+import com.forever.dadamda.dto.scrap.other.GetOtherResponse;
+import com.forever.dadamda.dto.scrap.product.GetProductCountResponse;
+import com.forever.dadamda.dto.scrap.product.GetProductResponse;
 import com.forever.dadamda.dto.scrap.GetScrapCountResponse;
 import com.forever.dadamda.dto.scrap.GetScrapResponse;
-import com.forever.dadamda.dto.scrap.GetVideoCountResponse;
-import com.forever.dadamda.dto.scrap.GetVideoResponse;
+import com.forever.dadamda.dto.scrap.video.GetVideoCountResponse;
+import com.forever.dadamda.dto.scrap.video.GetVideoResponse;
 import com.forever.dadamda.dto.scrap.UpdateScrapRequest;
 import com.forever.dadamda.entity.scrap.Scrap;
 import com.forever.dadamda.service.MemoService;
