+++ src/main/java/com/forever/dadamda/dto/scrap/article/GetArticleResponse.java
@@ -1,8 +1,8 @@
-package com.forever.dadamda.dto.scrap;
+package com.forever.dadamda.dto.scrap.article;
 
+import com.forever.dadamda.dto.memo.GetMemoResponse;
 import com.forever.dadamda.entity.scrap.Article;
 import com.forever.dadamda.service.TimeService;
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.AccessLevel;
