+++ src/main/java/com/forever/dadamda/dto/scrap/other/GetOtherCountResponse.java
@@ -1,11 +1,11 @@
-package com.forever.dadamda.dto.scrap;
+package com.forever.dadamda.dto.scrap.other;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor(staticName = "of")
-public class GetProductCountResponse {
+public class GetOtherCountResponse {
 
     private Long count;
 }
