+++ src/main/java/com/forever/dadamda/dto/scrap/product/GetProductCountResponse.java
@@ -1,12 +1,11 @@
-package com.forever.dadamda.dto.scrap;
+package com.forever.dadamda.dto.scrap.product;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor(staticName = "of")
-public class GetArticleCountResponse {
+public class GetProductCountResponse {
 
     private Long count;
-
 }
