+++ src/main/java/com/forever/dadamda/dto/scrap/video/GetVideoCountResponse.java
@@ -1,11 +1,11 @@
-package com.forever.dadamda.dto.scrap;
+package com.forever.dadamda.dto.scrap.video;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor(staticName = "of")
-public class GetOtherCountResponse {
+public class GetVideoCountResponse {
 
     private Long count;
 }
