+++ src/test/java/com/forever/dadamda/service/ScrapServiceTest.java
@@ -8,7 +8,7 @@ import com.forever.dadamda.entity.user.Provider;
 import com.forever.dadamda.entity.user.Role;
 import com.forever.dadamda.entity.user.User;
 import com.forever.dadamda.exception.NotFoundException;
-import com.forever.dadamda.repository.ScrapRepository;
+import com.forever.dadamda.repository.scrap.ScrapRepository;
 import com.forever.dadamda.repository.UserRepository;
 import com.forever.dadamda.service.scrap.ScrapService;
 import org.junit.jupiter.api.AfterEach;
